dr-monster-api-from-huggingface/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
Pillow
python-multipart
huggingface_hub